OBMFULL-3305 Send CANCEL itip for negative exceptions.
OBM-UI now sends a CANCEL itip upon removing an occurrence of a recurrent event.

Please note that this patch is an abomination. The mechanism for sending an itip
involves converting to an intermediate format before generating the ICS
file. The converter fetches the data from the database, with no knowledge of
the operation of suppression which just took place. The patch adds a
recurrenceId parameter, which forces the date of each retrieved event to be set
to this recurrenceId, and adds a recurrence ID property to these events. This
lets the ICS generator create a correct CANCEL itip message.