OBMFULL-3530 UI->obm-sync auth with md5 passwords.
Adds an optional 'isAlreadyHashed' parameter to the call to obm-sync's login
handler. When set to true, passwords with a type of 'MD5SUM' in the database are
compared directly against their value from the request, instead of hashing this
value. This lets both the UI (which sends hashed passwords in this case) and the
connector (which does not) work together.