OBMFULL-5876 RecurrenceKind computation refactor.
Put the RecurrenceKind value in its own variable, and assigns it in a single
location. This will allow us to rely on its value in other methods more
robustly, without having to worry about the possibility that it hasn't been
assigned yet in Event.

Port of OBMFULL-4633 to next branch.