OBMFULL-6022 Changes _userpattern.pattern to TEXT, improves index.
The change to TEXT is motivated by the fact that with PG 9.4+, it seems to
result in performance improvements in some setups (which is quite surprising).
The addition of text_pattern_ops to the index makes LIKE operations with text
anchored at the beginning much faster.

Port of OBMFULL-5923 to next branch.